requirements.txt
fastapi
uvicorn[standard]
elastic-opentelemetry

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
